+++ src/Settings.php
@@ -113,14 +113,14 @@ class Settings {
 	private $callback;
 
 	/**
-	 * The used URL.
+	 * The used URL of this composer package.
 	 *
 	 * @var string
 	 */
 	private string $url = '';
 
 	/**
-	 * The used path.
+	 * The used path of this composer package.
 	 *
 	 * @var string
 	 */
@@ -148,13 +148,6 @@ class Settings {
 	private array $translations = array();
 
 	/**
-	 * The default styling.
-	 *
-	 * @var string
-	 */
-	private string $styling = 'horizontal_tabs';
-
-	/**
 	 * The import object.
 	 *
 	 * @var Import
@@ -167,6 +160,13 @@ class Settings {
 	 * @var Export
 	 */
 	private Export $export_obj;
+
+	/**
+	 * The views object.
+	 *
+	 * @var Views
+	 */
+	private Views $views;
 
 	/**
 	 * Constructor, not used as this a Singleton object.
@@ -185,6 +185,9 @@ class Settings {
 
 			// prepare the methods.
 			Methods::get_instance()->set_settings_obj( $this );
+
+			// prepare the views.
+			$this->views = new Views( $this );
 		} catch ( \Exception $e ) {
 			return;
 		}
@@ -217,7 +220,6 @@ class Settings {
 		// use hooks.
 		add_action( 'admin_menu', array( $this, 'add_menu' ) );
 		add_action( 'admin_init', array( $this, 'register_fields' ) );
-		add_action( 'admin_enqueue_scripts', array( $this, 'add_js_and_css' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'add_dialog' ) );
 
 		// use our own hooks.
@@ -506,21 +508,21 @@ class Settings {
 	}
 
 	/**
+	 * Return the views object.
+	 *
+	 * @return Views
+	 */
+	public function get_views(): Views {
+		return $this->views;
+	}
+
+	/**
 	 * Show the navigation of settings.
 	 *
 	 * @return void
 	 */
 	public function display(): void {
-		// get the styling object.
-		$styling_object = $this->get_styling_object();
-
-		// bail if no styling object could be found.
-		if ( ! $styling_object instanceof Styling_Base ) {
-			return;
-		}
-
-		// show the navigation.
-		$styling_object->show_nav();
+		$this->get_views()->display();
 	}
 
 	/**
@@ -927,7 +929,7 @@ class Settings {
 	}
 
 	/**
-	 * Return the used URL.
+	 * Return the used URL for this composer package.
 	 *
 	 * @return string
 	 */
@@ -1035,72 +1037,7 @@ class Settings {
 	}
 
 	/**
-	 * Add own JS and CSS for the backend.
-	 *
-	 * @param string $hook The requested hook.
-	 * @return void
-	 */
-	public function add_js_and_css( string $hook ): void {
-		// bail if not the menu slug is called.
-		if ( ! $this->enqueue_styles_and_scripts( $hook ) ) {
-			return;
-		}
-
-		// add backend JS.
-		wp_enqueue_script(
-			$this->get_slug() . '-settings',
-			$this->get_url() . 'assets/js.js',
-			array( 'jquery', 'jquery-ui-sortable', 'jquery-ui-draggable', 'jquery-ui-droppable' ),
-			Helper::get_file_version( $this->get_path() . 'assets/js.js', $this ),
-			true
-		);
-
-		// add dirty.js.
-		wp_enqueue_script(
-			$this->get_slug() . '-dirty',
-			$this->get_url() . 'assets/jquery.dirty.js',
-			array( 'jquery' ),
-			Helper::get_file_version( $this->get_path() . 'assets/jquery.dirty.js', $this ),
-			true
-		);
-
-		// add backend CSS.
-		wp_enqueue_style(
-			$this->get_slug() . '-settings',
-			$this->get_url() . 'assets/style.css',
-			array(),
-			Helper::get_file_version( $this->get_path() . 'assets/style.css', $this ),
-		);
-
-		// add CSS for chosen styling.
-		$styling_object = $this->get_styling_object();
-		if ( $styling_object instanceof Styling_Base ) {
-			$styling_object->add_styles();
-		}
-
-		// get the translations.
-		$translations = $this->get_translations();
-
-		// add php-vars to our js-script.
-		wp_localize_script(
-			$this->get_slug() . '-settings',
-			'esfwJsVars',
-			array(
-				'rest_settings'        => rest_url( 'wp/v2/settings' ),
-				'rest_nonce'           => wp_create_nonce( 'wp_rest' ),
-				'title_add_image'      => $translations['file_add_file'],
-				'button_add_image'     => $translations['file_choose_file'],
-				'lbl_upload_image'     => $translations['file_choose_image'],
-				'label_sortable_title' => $translations['drag_n_drop'],
-			)
-		);
-
-		// add media library.
-		wp_enqueue_media();
-	}
-
-	/**
-	 * Return the path.
+	 * Return the path to this composer package.
 	 *
 	 * It has a trailing flash.
 	 *
@@ -1388,84 +1325,6 @@ class Settings {
 	}
 
 	/**
-	 * Return the object of the configured styling.
-	 *
-	 * @return Styling_Base|false
-	 */
-	public function get_styling_object(): Styling_Base|false {
-		// prepare the result.
-		$style_obj = false;
-
-		// check each supported styling for the configured styling name.
-		foreach ( $this->get_styling_objects() as $styling_name ) {
-			// bail if the class name does not exist.
-			if ( ! class_exists( $styling_name ) ) {
-				continue;
-			}
-
-			// get the object.
-			$obj = new $styling_name( $this );
-
-			// bail if an object is not Schedules_Base.
-			if ( ! $obj instanceof Styling_Base ) {
-				continue;
-			}
-
-			// bail if name does not match.
-			if ( $obj->get_name() !== $this->get_styling() ) {
-				continue;
-			}
-
-			// use this object.
-			$style_obj = $obj;
-		}
-
-		// return the resulting object.
-		return $style_obj;
-	}
-
-	/**
-	 * Return the configured styling name.
-	 *
-	 * @return string
-	 */
-	private function get_styling(): string {
-		return $this->styling;
-	}
-
-	/**
-	 * Set the styling to use by its name.
-	 *
-	 * @param string $styling The styling name.
-	 *
-	 * @return void
-	 * @noinspection PhpUnused
-	 */
-	public function set_styling( string $styling ): void {
-		$this->styling = $styling;
-	}
-
-	/**
-	 * Return the list of possible styling objects.
-	 *
-	 * @return array<int,string>
-	 */
-	private function get_styling_objects(): array {
-		$list = array(
-			'\easySettingsForWordPress\Styles\Horizontal_Tabs',
-			'\easySettingsForWordPress\Styles\Vertical_Tabs',
-		);
-
-		/**
-		 * Filter the list of possible styling for settings.
-		 *
-		 * @since 2.0.0 Available since 2.0.0.
-		 * @param array $list The list.
-		 */
-		return apply_filters( $this->get_slug() . '_styling_objects', $list );
-	}
-
-	/**
 	 * Return the used settings for debug purposes.
 	 *
 	 * @return array<string,mixed>
@@ -1478,11 +1337,18 @@ class Settings {
 			$method_name = $method->get_name();
 		}
 
+		// get the view.
+		$view = $this->views->get_view();
+		$view_name = '';
+		if( $view instanceof View_Base ) {
+			$view_name = $view->get_name();
+		}
+
 		// return the settings.
 		return array(
 			'plugin_path' => $this->get_plugin_path(),
 			'method'      => $method_name,
-			'styling'     => $this->get_styling(),
+			'view'        => $view_name,
 			'settings'    => $this->get_settings(),
 		);
 	}
@@ -1498,8 +1364,31 @@ class Settings {
 	 * @param string $old_method_name The name of the previous method.
 	 *
 	 * @return void
+	 * @noinspection PhpUnused
 	 */
 	public function migrate_method( string $old_method_name ): void {
 		Methods::get_instance()->migrate_method( $old_method_name );
+	}
+
+	/**
+	 * Set the method to use.
+	 *
+	 * @param string $method_name The method name.
+	 *
+	 * @return void
+	 */
+	public function set_method( string $method_name ): void {
+		Methods::get_instance()->set_method( $method_name );
+	}
+
+	/**
+	 * Set the view to use.
+	 *
+	 * @param string $view_name The method name.
+	 *
+	 * @return void
+	 */
+	public function set_view( string $view_name ): void {
+		$this->views->set_view( $view_name );
 	}
 }